Refactor how the ldap server parameters are built
Instead of using a crazy nest of inline stdlib functions create and use a
custom function that when given an array of ldap(s):/// URI strings will
return an array of parameters suitable to merge into the login class
definition.

By switching to using one or more URI instead of just bare hostnames or IP
addresses the rest of the connection details can be conveyed without needing
complex structures or multiple parameters.